#90: refactor: extract shared linkedin XML validation into model.mjs
Pulls parseXml/validateDocument out of validate.mjs into an
environment-agnostic model.mjs so the same validation rules can be
reused from the browser-side editor UI added for #90.
